requirements.txt
fastapi==0.115.8
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
SQLAlchemy==2.0.37
uvicorn==0.34.0
